Moving storage connection default logic into _implicit_environ.
