+++ gcloud/storage/_implicit_environ.py
@@ -19,6 +19,15 @@ from the enviroment.
 """
 
 
+from gcloud.connection import get_scoped_connection
+from gcloud.storage.connection import Connection
+
+
+SCOPE = ('https://www.googleapis.com/auth/devstorage.full_control',
+         'https://www.googleapis.com/auth/devstorage.read_only',
+         'https://www.googleapis.com/auth/devstorage.read_write')
+
+
 class _DefaultsContainer(object):
     """Container for defaults.
 
@@ -50,6 +59,33 @@ def get_default_connection():
     :returns: The default connection if one has been set.
     """
     return _DEFAULTS.connection
+
+
+def get_connection():
+    """Shortcut method to establish a connection to Cloud Storage.
+
+    Use this if you are going to access several buckets with the same
+    set of credentials:
+
+    >>> from gcloud import storage
+    >>> connection = storage.get_connection()
+    >>> bucket1 = storage.get_bucket('bucket1', connection=connection)
+    >>> bucket2 = storage.get_bucket('bucket2', connection=connection)
+
+    :rtype: :class:`gcloud.storage.connection.Connection`
+    :returns: A connection defined with the proper credentials.
+    """
+    return get_scoped_connection(Connection, SCOPE)
+
+
+def set_default_connection(connection=None):
+    """Set default connection either explicitly or implicitly as fall-back.
+
+    :type connection: :class:`gcloud.storage.connection.Connection`
+    :param connection: A connection provided to be the default.
+    """
+    connection = connection or get_connection()
+    _DEFAULTS.connection = connection
 
 
 _DEFAULTS = _DefaultsContainer()
